# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "EICE API",
    "description": "Elpo Interview Coaching Experts API - Candidates and Questions",
    "version": "1.0.0"
  },
  "host": "eice-api.onrender.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "sessionCookie": {
      "type": "apiKey",
      "name": "Cookie",
      "in": "header",
      "description": "Authenticate at /login first. Send the resulting connect.sid session cookie."
    }
  },
  "paths": {
...
    "/github/callback": {
      "get": {
        "description": "GitHub OAuth callback. On success, redirects to Swagger UI with an authenticated session.",
        "responses": {
          "302": {
            "description": "Redirects to Swagger UI"
          },
          "401": {
            "description": "Authentication failed"
          }
        }
      }
    },
...
    "/auth/logout": {
      "get": {
        "description": "Logs out the current user and destroys the session.",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logged out successfully"
          },
          "500": {
            "description": "Logout failed"
          }
        }
      }
    },
...
  }
}